backend/requirements.txt
Flask
pandas
gunicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
